Add files via upload
How it works- Open the mod and send a trade request to another player by pressing the "Send Trade" button and then typing their name. Next select the service you are willing to give the other player. Hit accept to finish the trade. If you fail to give the player the service they have demanded, you will be put on a "Scammer" list which will be uploaded to multiple different databases and discords. This is a new mod so expect bugs Send ideas to W.O.T#3162 (Discord)

I recommend using the Skyblock Trading Menu as well as this so that players can trade money/items at the same time as a service (Yes there is a feature for this).